Add output versioning.
Output is now versioned with a ".p{NUM}." infix, i.e., penultimate
suffix. This versioning is applied for restarts as well as reruns
as long as the output-prefix is not changed.

Further, output is now managed via an explicit output module that holds
specific output writer classes for the existing xyz- and vtu-output as
well as for a new xdmf writer with hdf5 storage.