Use OmpGuard as a scoped lock.
As a leftover from the flip development, the OmpGuard was reduced
to just wrap omp_test_lock and omp_unset_lock. All handling
of not leaving locked 'locks' around had to be done in the flip
routine itself. This is now improved by making use of the
OmpGuard as the intended scoped lock.