Improve config-file documentation.
The 'config' subcommand of the mc_app cli now by default shows
a verbosely commented configuration. It can be stripped with the
'--strip' flag and written to a file with the '--conf' flag. The
latter is as before but not mandatory any longer. By default the
configuration is printed to screen.