Remove atomics from gradient evaluation.
The atomic operations have been identified to take up a
significant amount of time in the gradient evaluation. They can be
entirely avoided by not accumulating gradients for each vertex
involved in a patch (of vertex-averaged properties) but by
accumulating for each vertex, the gradients of its contributions
to the vertex properties of patches were it is involved in, ie.,
its own as well as the neighbouring ones. This also simplifies the
evaluation of the gradients of mesh properties, since now
the symmetries therein can be exploited. That is, cyclic-symmetry
for face-area and face-volume wrt a halfedge cycle and
symmetry wrt an exchange of a halfedge with its opposite for
edge length and dihedral angle.